Move messenger/ to chat_service/ (#2098)
* Move messenger/ to chat_service/

* Fix unused imports

* Move all messenger code to chat_service/services/messenger

* Update all pointers to directory structure
