+++ parlai/chat_service/tasks/chatbot/worlds.py
@@ -4,10 +4,10 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 #
-# py parlai/messenger/tasks/overworld_demo/run.py --debug --verbose
+# py parlai/chat_service/tasks/overworld_demo/run.py --debug --verbose
 
 from parlai.core.worlds import World
-from parlai.messenger.core.worlds import OnboardWorld
+from parlai.chat_service.services.messenger.worlds import OnboardWorld
 from parlai.core.agents import create_agent_from_shared
