+++ parlai/chat_service/tasks/qa_data_collection/worlds.py
@@ -3,10 +3,6 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-from parlai.messenger.core.worlds import (
-    SimpleMessengerOverworld as MessengerOverworld,
-    OnboardWorld as QADataCollectionOnboardWorld,
-)
 from parlai.core.worlds import World, validate
 import importlib
